staff profile page implemented, some locale updates, fixed the dasbhoard redirection

+++ src/interceptor.js
@@ -18,16 +18,16 @@ export default function setup() {
         return Promise.reject(err);
     });
     axios.interceptors.response.use(function (response) {
-        if (response.status == 401) {
-            userService.removeLoginInfo();
-            router.push({ name: "Login Dashboard" })
-        }
+        checkAccess(response);
         return response;
     }, function (error) {
-        if (error.response.status == 401) {
-            userService.removeLoginInfo();
-            router.push({ name: "Login Dashboard" })
-        }
+        checkAccess(error.response);
         return Promise.reject(error);
     })
+}
+function checkAccess(response) {
+    if (response && response.status == 401 && !router.currentRoute.path.includes('auth')) {
+        userService.removeLoginInfo();
+        router.push({ name: "Login Dashboard" })
+    }
 }
